@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.5.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -22,15 +22,32 @@ dependencies:
   http: ^0.12.0+4
   charts_flutter: ^0.9.0
   random_color: ^1.0.5
-  flutter_statusbarcolor: ^0.2.3
+  google_fonts: ^1.1.0
+  sticky_headers: ^0.1.8+1
+  url_launcher: ^5.4.7
+  share: ^0.6.4+2
+  fluttertoast: ^4.0.1
+  font_awesome_flutter: ^8.8.1
+  # sqflite: ^1.3.0+1
+  flutter_screenutil: ^1.1.0
+  hive: ^1.4.1+1
+  hive_flutter: ^0.3.0+1
+  cupertino_icons: ^0.1.3    
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+# dependency_overrides:
+#   dartx: ^0.3.0  
+#   path: ^1.7.0
+#   quiver: 2.1.0
 
 dev_dependencies:
+  flutter_driver:
+    sdk: flutter
+  test: any
   flutter_test:
     sdk: flutter
+  hive_generator: 0.5.1
+  build_runner: ^1.10.0
+  
 
 
 # For information on the generic Dart part of this file, see the
