@@ -1,6 +1,9 @@
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+// import 'package:stocks_app/data/stockDayDataModel.dart';
+import 'package:stocks_app/data/stock_day.dart';
 
 class StockSeries {
   final String date;
@@ -11,16 +14,16 @@ class StockSeries {
 
 class StockChart extends StatelessWidget {
   final List<StockSeries> dataChart;
-  final Map dataMapped;
+  final StockDayModel dataMapped;
 
-  StockChart({@required this.dataChart,@required this.dataMapped});
+  StockChart({@required this.dataChart, @required this.dataMapped});
 
   @override
   Widget build(BuildContext context) {
     List<charts.Series<StockSeries, DateTime>> series = [
       charts.Series(
           id: "Stocks",
-          colorFn: (_, __) => charts.MaterialPalette.pink.shadeDefault,
+          colorFn: (_, __) => charts.MaterialPalette.blue.shadeDefault,
           data: dataChart,
           domainFn: (StockSeries series, _) => DateTime(
               int.parse(series.date.split("-")[0]),
@@ -31,21 +34,23 @@ class StockChart extends StatelessWidget {
     ];
     return Container(
       color: Colors.white,
-      height: 400,
+      height: 650.h,
       child: Container(
         child: Padding(
           padding: const EdgeInsets.all(8.0),
           child: Column(
             children: <Widget>[
-              Text(this.dataMapped["close"].toString(),
-                style: TextStyle(fontWeight: FontWeight.w800,fontSize: 50),
+              Text(
+                this.dataMapped.close.toString(),
+                style: TextStyle(fontWeight: FontWeight.w800, fontSize: 50),
               ),
               Expanded(
                 child: charts.TimeSeriesChart(
                   series,
                   animate: true,
                   animationDuration: Duration(milliseconds: 500),
-                  defaultRenderer:new charts.LineRendererConfig(includeArea: true, stacked: true),
+                  defaultRenderer: new charts.LineRendererConfig(
+                      includeArea: true, stacked: true),
                 ),
               )
             ],
