@@ -1,6 +1,6 @@
-import 'package:flutter_statusbarcolor/flutter_statusbarcolor.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/material.dart';
-import 'home.dart';
+import 'package:stocks_app/screens/home.dart';
 
 void main() => runApp(MyApp());
 
@@ -15,8 +15,8 @@ class MyBehavior extends ScrollBehavior {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    FlutterStatusbarcolor.setStatusBarColor(Colors.white);
-    FlutterStatusbarcolor.setStatusBarWhiteForeground(false);
+    SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(statusBarColor: Colors.white));
     return MaterialApp(
       builder: (context, child) {
         return ScrollConfiguration(
@@ -27,7 +27,8 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Stock Market',
       theme: ThemeData(
-        primarySwatch: Colors.pink,
+        appBarTheme: AppBarTheme(brightness: Brightness.light),
+        primarySwatch: Colors.blue,
       ),
       home: Home(title: 'Stock Market'),
     );
